requirements.txt
numpy==1.20.0
cython
PyOpenGL-accelerate
hydra-core
protobuf==3.20.1
gymnasium
dm_control
opencv-python
kornia
imageio==2.31.5
imageio[ffmpeg]
imageio[pyav]
ray==1.0.0
torchrl==0.1.0
